Finalise restructuring from DAO to Safe prefixed components
