--- src/pages/daos/[daoAddress]/proposal-templates/index.tsx
@@ -1,77 +1,0 @@
-import * as amplitude from '@amplitude/analytics-browser';
-import { Box, Button, Flex, Show } from '@chakra-ui/react';
-import { useEffect } from 'react';
-import { useTranslation } from 'react-i18next';
-import { Link } from 'react-router-dom';
-import { AddPlus } from '../../../../assets/theme/custom/icons/AddPlus';
-import ProposalTemplateCard from '../../../../components/ProposalTemplates/ProposalTemplateCard';
-import NoDataCard from '../../../../components/ui/containers/NoDataCard';
-import { InfoBoxLoader } from '../../../../components/ui/loaders/InfoBoxLoader';
-import PageHeader from '../../../../components/ui/page/Header/PageHeader';
-import { DAO_ROUTES } from '../../../../constants/routes';
-import { useCanUserCreateProposal } from '../../../../hooks/utils/useCanUserSubmitProposal';
-import { analyticsEvents } from '../../../../insights/analyticsEvents';
-import { useFractal } from '../../../../providers/App/AppProvider';
-import { useNetworkConfig } from '../../../../providers/NetworkConfig/NetworkConfigProvider';
-
-export default function ProposalTemplatesPage() {
-  useEffect(() => {
-    amplitude.track(analyticsEvents.ProposalTemplatesPageOpened);
-  }, []);
-
-  const { t } = useTranslation();
-  const {
-    node: { daoAddress },
-    governance: { proposalTemplates },
-  } = useFractal();
-  const { canUserCreateProposal } = useCanUserCreateProposal();
-  const { addressPrefix } = useNetworkConfig();
-
-  return (
-    <div>
-      <PageHeader
-        title={t('proposalTemplates', { ns: 'breadcrumbs' })}
-        breadcrumbs={[
-          {
-            terminus: t('proposalTemplates', { ns: 'breadcrumbs' }),
-            path: '',
-          },
-        ]}
-      >
-        {canUserCreateProposal && daoAddress && (
-          <Link to={DAO_ROUTES.proposalTemplateNew.relative(addressPrefix, daoAddress)}>
-            <Button minW={0}>
-              <AddPlus />
-              <Show above="sm">{t('create')}</Show>
-            </Button>
-          </Link>
-        )}
-      </PageHeader>
-      <Flex
-        flexDirection={proposalTemplates && proposalTemplates.length > 0 ? 'row' : 'column'}
-        flexWrap="wrap"
-        gap="1rem"
-      >
-        {!proposalTemplates ? (
-          <Box>
-            <InfoBoxLoader />
-          </Box>
-        ) : proposalTemplates.length > 0 ? (
-          proposalTemplates.map((proposalTemplate, i) => (
-            <ProposalTemplateCard
-              key={i}
-              proposalTemplate={proposalTemplate}
-              templateIndex={i}
-            />
-          ))
-        ) : (
-          <NoDataCard
-            translationNameSpace="proposalTemplate"
-            emptyText="emptyProposalTemplates"
-            emptyTextNotProposer="emptyProposalTemplatesNotProposer"
-          />
-        )}
-      </Flex>
-    </div>
-  );
-}
